requirements.txt
# HSE Vision - Professional Helmet Detection System Requirements

# Core ML/CV Libraries
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0

# Data Processing
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# GUI Framework
tkinter-tooltip>=2.0.0
customtkinter>=5.2.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
pathlib2>=2.3.7

# Development Tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For advanced features
# tensorrt>=8.6.0  # For model optimization (NVIDIA GPUs only)
# onnx>=1.14.0     # For model conversion
# wandb>=0.15.0    # For experiment tracking